[BinderTransport] add authority to initial metadata if missing
When C++ implementation acts as server and Java implementation acts as
client, authority will not be present in initial metadata. Although
authority is not used for binder transport, when it is missing, core API
layer will fail the RPC call silently.

This commit adds a fake authority to initial metadata when it is
missing to avoid the failure.

Test case is added to make sure the wire reader implementation appends
both path and authority to metadata correctly.